new idle

+++ src/gameworld/PRJ_Control.java
@@ -126,8 +126,8 @@ public class PRJ_Control {
         if (entity.health <= 0) {
             mg.player.getExperience(entity);
             entity.dead = true;
-            mg.WORLD_DROPS.add(new DRP_DroppedItem(mg, entity.worldX, entity.worldY, entity.level));
-            mg.WORLD_DROPS.add(new DRP_Coin(entity.worldX + mg.random.nextInt(41) - 20, entity.worldY + mg.random.nextInt(41) - 20, entity.level));
+            mg.WORLD_DROPS.add(new DRP_DroppedItem(mg, (int) entity.worldX, (int) entity.worldY, entity.level));
+            mg.WORLD_DROPS.add(new DRP_Coin((int) (entity.worldX + mg.random.nextInt(41) - 20), (int) (entity.worldY + mg.random.nextInt(41) - 20), entity.level));
         } else {
             entity.hpBarOn = true;
         }
