+++ src/main/system/ui/maps/MiniMap.java
@@ -41,8 +41,8 @@ public class MiniMap {
         synchronized (mg.PROXIMITY_ENTITIES) {
             gc.setFill(Colors.Red);
             for (gameworld.entities.ENTITY entity : mg.PROXIMITY_ENTITIES) {
-                entityX = (entity.worldX + 24) / 48;
-                entityY = (entity.worldY + 24) / 48;
+                entityX = (int) ((entity.worldX + 24) / 48);
+                entityY = (int) ((entity.worldY + 24) / 48);
                 if ((entityX - xTile) < 20 && xTile - entityX <= 20 && (entityY - yTile) < 20 && yTile - entityY <= 20 && !(entity instanceof ENT_Owly)) {
                     gc.fillRect(1_700 + 100 + (entityX - xTile) * 5, 25 + 100 + (entityY - yTile) * 5, 5, 5);
                 }
@@ -61,8 +61,8 @@ public class MiniMap {
         }
         gc.setFill(Colors.blue_npc);
         for (ENTITY entity : mg.npcControl.NPC_Active) {
-            entityX = (entity.worldX + 24) / 48;
-            entityY = (24 + entity.worldY) / 48;
+            entityX = (int) ((entity.worldX + 24) / 48);
+            entityY = (int) ((24 + entity.worldY) / 48);
 
             gc.fillRect(1_700 + 100 + (entityX - xTile) * 5, 25 + 100 + (entityY - yTile) * 5, 5, 5);
         }
